% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.PARAFAC-Log.R
\name{Biplot.PARAFACLog}
\alias{Biplot.PARAFACLog}
\title{Biplot for PARAFAC-Logit results}
\usage{
Biplot.PARAFACLog(PARLog)
}
\arguments{
\item{PARLog}{A list containing the results of a PARAFAC-Logit model,
including factor matrices, predictions, and other relevant data.}
}
\value{
A list containing the biplot information, including row and column
coordinates, scale factor, eigenvalues, inertia, contributions, and various
fit statistics.
}
\description{
This function generates a biplot from the results of a PARAFAC-Logit model,
including various statistics and contributions.
}
\examples{
PARLog <- PARAFAC_Logit (X=X3way, Y=Y, I=8, J=7, K=6, L=9)
Biplot<-Biplot.PARAFACLog(PARLog)
inames = paste("i",1:8,sep="")
xnames=NULL
for (k in 1:6)
      for (j in 1:7){
        xnames=c(xnames, paste(k, j,sep="."))
      }
plot(Biplot,IndLabels =inames,VarLabels=xnames)
}
