% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAFAC_Log.R
\name{PARAFAC_Logit}
\alias{PARAFAC_Logit}
\title{This function fits a PARAFAC-Log model to a coupled data set consisting of
a three-way array of real data and a binary data matrix that share a mode.}
\usage{
PARAFAC_Logit(
  X,
  Y,
  I,
  J,
  K,
  L,
  inames = NULL,
  xnames = NULL,
  ynames = NULL,
  R = 2,
  alfa = 0.5,
  tolerance = 5e-06,
  maxiter = 100,
  penalization = 0.1,
  OptimMethod = "CG"
)
}
\arguments{
\item{X}{The first mode unfolding of a three-way array of real-valued data.}

\item{Y}{A binary data matrix.}

\item{I}{Number of elements in the first mode in X.}

\item{J}{Number of elements in the second mode in X.}

\item{K}{Number of elements in the third mode in X.}

\item{L}{Number of columns in Y.}

\item{R}{Number of components in the PARAFAC model.}

\item{alfa}{Weighting parameter.}

\item{tolerance}{Convergence tolerance.}

\item{maxiter}{Maximum number of iterations.}

\item{penalization}{Learning rate,}

\item{OptimMethod}{Optimization method to use (default is "CG").}

\item{inames:}{Optional, names for the elements in the first mode of X.}

\item{xnames:}{Optional, names for the elements in the second/third mode of X.}

\item{ynames:}{Optional, names of the variables of Y.}
}
\value{
A list with the following elements:
\item{A}{Component matrix A.}
\item{B1}{Component matrix B1.}
\item{C}{Component matrix C.}
\item{B2}{Factor matrix B2.}
\item{b0}{Intercept.}
\item{L}{Final value of the loss function.}
\item{Expected}{Matrix of expected values.}
\item{Predictions}{Matrix of binary predictions.}
}
\description{
This function fits a PARAFAC-Log model to a coupled data set consisting of
a three-way array of real data and a binary data matrix that share a mode.
}
\examples{
PARLog <- PARAFAC_Logit (X=X3way, Y=Y, I=8, J=7, K=6, L=9)
}
