\name{Y}
\alias{Y}
\docType{data}
\title{Two way matrix}
\description{Gathers information about whether or not a group of 8 people have certain personality traits.}
\usage{data(Y)}
\format{
Dispositions are the following:
    \cr{Fear to be refused}
    \cr{Kindness}
    \cr{Importance of others' judgments}
    \cr{Altruism}
    \cr{Neuroticism}
    \cr{Being strict to oneself}
    \cr{Low selfesteem}
    \cr{Conscientiousness}
    \cr{Depression}
  }
\examples{
data(Y)
str(Y)
}
\keyword{datasets}
